bpf: Add a new function
The netif_receive_generic-xdp function
This function is the entry point for processing inbound packets. It attempts to execute an XDP program (xdp0pro) and determines the subsequent processing path of the packet based on the program's return value
Generic_xdp_tx function
This function is used to process data packets that XDP programs decide to send.
Do-xdp_generic function
This function is the entry point for general XDP processing, which determines how to process data packets based on the execution results of the XDP program.